Implement filtering todos by owner
This commit adds the necessary code to filter the todos by their owner, with owner being filtered on the server. This commit also fixes the test in todo.service.spec.ts that tests to see if getTodos() correctly calls api/todos with filter parameter of owner.  The wrong parameter type was given in the test, so the test failed. The parameter type was changed to the correct one and now the test passes. Issue #20