Implement todo-info component
This feature implements todo-info component. Now, users can view specific todo information by doing "/api/todos/:id". This idea is used in addTodo (which has yet to be implemented) in the sense that when users add a new todo, they will be brought to a page that displays the information of the todo that they just added. Issue #11 Issue #23