Fix errors and redesign tests for add todo
This commit fixes a error in add-todo.po.ts in which an instance of "name" was found instead of "owner". This commit also redesigns the tests so that addTodo is more like addUser in the sense one of the fields is not required by angular, but is required by the server. In the case of todos, category is the field that is not required by angular but is required by the server.