Change Instances of Status Boolean Tests
This converts the status html from showing a boolean to showing 'complete' or 
'incomplete'. This also changes the allowed status input for a new todo with the options above. Add labels for data in html in todo list component. Unit and E2E tests for status portions are changed to accommodate for these display changes.
co-authored-by: Jeklund99 <eklun124@umn.edu>